% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_category.R
\name{plot_region_category}
\alias{plot_region_category}
\title{Plot Number of Heritage Sites by Region and Category}
\usage{
plot_region_category(data)
}
\arguments{
\item{heritage_clean}{The cleaned heritage dataset.}
}
\value{
A ggplot object showing the distribution of sites by region and category.
}
\description{
This function creates a horizontal bar plot showing the number of World Heritage Sites by region and category.
}
