% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trends.R
\name{plot_trends}
\alias{plot_trends}
\title{Plot Trends in World Heritage Site Inscriptions by Region}
\usage{
plot_trends(data)
}
\arguments{
\item{heritage_clean}{The cleaned heritage dataset.}
}
\value{
A ggplot object showing the trends.
}
\description{
This function creates a time series plot showing the number of World Heritage Sites inscribed each year by region.
}
