% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heritage.R
\name{print.heritage}
\alias{print.heritage}
\title{Print a fitted HERITAGE model}
\usage{
\method{print}{heritage}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"heritage"}.}

\item{...}{Currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a fitted HERITAGE model
}
