% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heritage_gs.R
\name{print.heritage_gs}
\alias{print.heritage_gs}
\title{Print grid-search results for HERITAGE}
\usage{
\method{print}{heritage_gs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"heritage_gs"}.}

\item{...}{Currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print grid-search results for HERITAGE
}
