% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heritage.R
\name{summary.heritage}
\alias{summary.heritage}
\title{Summarise a fitted HERITAGE model}
\usage{
\method{summary}{heritage}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"heritage"}.}

\item{...}{Currently unused.}
}
\value{
\code{object}, invisibly.
}
\description{
Prints a detailed summary including all non-zero coefficients.
}
