% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heritage_gs.R
\name{summary.heritage_gs}
\alias{summary.heritage_gs}
\title{Summarise grid-search results for HERITAGE}
\usage{
\method{summary}{heritage_gs}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"heritage_gs"}.}

\item{...}{Currently unused.}
}
\value{
\code{object}, invisibly.
}
\description{
Summarise grid-search results for HERITAGE
}
