% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_top_countries.R
\name{table_top_countries}
\alias{table_top_countries}
\title{Generate a Data Frame of Top 10 Countries with the Most World Heritage Sites}
\usage{
table_top_countries(data)
}
\arguments{
\item{data}{The cleaned heritage dataset.}
}
\value{
A data frame of top countries.
}
\description{
This function generates a data frame showing the top 10 countries with the highest number of World Heritage Sites.
}
